refactor: Game 클래스 삭제 및 책임 분리
