+++ src/main/java/domain/Hands.java
@@ -1,6 +1,8 @@
 package domain;
 
 import domain.card.Card;
+import domain.result.Result;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
