+++ src/main/java/domain/participant/Players.java
@@ -1,10 +1,13 @@
 package domain.participant;
 
 import domain.Hands;
+import domain.result.Result;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 public class Players {
 
@@ -29,6 +32,14 @@ public class Players {
     public boolean isAllBust() {
         return names.stream()
                 .allMatch(Player::isBust);
+    }
+
+    public Map<Player, Result> getPlayersResult(final Dealer dealer) {
+        return names.stream()
+                .collect(Collectors.toMap(
+                        player -> player,
+                        player -> player.calculateResult(dealer)
+                ));
     }
 
     private static List<Player> mapToPlayers(final List<String> names) {
